requirements.txt
yfinance
pandas
matplotlib
seaborn
scikit-learn
newsapi-python
openai
python-dotenv
streamlit
requests